requirements.txt
annotated-types==0.6.0
anyio==4.2.0
black==23.12.1
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
fastapi==0.109.0
filelock==3.13.1
Flask==3.0.0
fsspec==2023.12.2
h11==0.14.0
httptools==0.6.1
huggingface-hub==0.20.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.3
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.1.0
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.1
sniffio==1.3.0
starlette==0.35.1
sympy==1.12
tokenizers==0.15.0
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tqdm==4.66.1
transformers==4.36.2
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.26.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
